fix: 修复初始化effact多次执行问题
